+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
+
+## 1.4.0 - 2021-08-02
 - Add `TracedError` type and module
 - Add `Trace.addError` function
 - Fix extract trace from http not to add tags, since it should be added on the referencing trace